Fix panics being shadowed by exit() in the defer statement
Running both defers.Recover() and exit() in the same defer, like so,
did not work:

    func Handle(a any) {
        var code int
        defer func() {
            defers.Recover()
            exit(code)
        }()
        // ... snip ...
    }

Despite the fact that this would have been nice to do to prevent any way
of exiting from the function without running exit() or panicking.

This is due to the way recover() works.

https://pkg.go.dev/builtin#recover

"Executing a call to recover inside a deferred function (but not any
function called by it) stops the panicking sequence by restoring normal
execution..."

Since defers.Recover() already executes defers in the event of a panic,
moving exit() out of the defers and ending the function with it should
still prevent any way of leaving Handle() without running all registered
defers.